requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
google-cloud-bigquery==3.25.0
google-cloud-bigquery-migration==0.14.0
google-cloud-aiplatform==1.66.0
tenacity==8.5.0
